requirements.txt
fastapi
uvicorn
python-dotenv
requests
prometheus-fastapi-instrumentator
loguru
